all needed pages done looks about ready to launch
