+++ frontend/app/chatbot/page.tsx
@@ -173,8 +173,8 @@ export default function Chatbot() {
                         <Image
                             src="/brand_logo_black.png"
                             alt="ConklinOfficial Logo"
-                            width={70}
-                            height={25}
+                            width={180}
+                            height={80}
                             className="object-contain"
                         />
                     </Link>
