+++ frontend/app/payment-success/page.tsx
@@ -25,8 +25,8 @@ export default function PaymentSuccessPage() {
             <Image 
               src="/brand_logo_black.png" 
               alt="ConklinOfficial Logo" 
-              width={70} 
-              height={25} 
+              width={180} 
+              height={80} 
               className="object-contain"
             />
           </Link>
