+++ frontend/app/product/[id]/page.tsx
@@ -69,8 +69,8 @@ export default function ProductDetail({ params }: { params: { id: string } }) {
             <Image
               src="/brand_logo_black.png"
               alt="ConklinOfficial Logo"
-              width={70}
-              height={25}
+              width={180}
+              height={80}
               className="object-contain"
             />
           </Link>
